+++ SchedulingProject/Jobs.h
@@ -11,20 +11,20 @@ public:
 	Jobs(vector<vector<ValType>> jobs);
 	~Jobs();
 
-	int serverCount();
-	int jobCount(int server);
+	size_t serverCount();
+	size_t jobCount(size_t server);
 
-	ValType getJob(int server, int index);
-	void setJob(int server, int index, ValType value);
+	ValType getJob(size_t server, size_t index);
+	void setJob(size_t server, size_t index, ValType value);
 
-	void split(Jobs &first, Jobs &remaining, int amount);
+	void split(Jobs &first, Jobs &remaining, size_t amount);
 	void add(Jobs &range);
 	ValType maxRunTime();
 
 	void addServer();
-	void addJob(int server, ValType length);
-	void removeServer(int server);
-	void removeJob(int server, int index);
+	void addJob(size_t server, ValType length);
+	void removeServer(size_t server);
+	void removeJob(size_t server, size_t index);
 
 	void jobsAfter(ValType syncPoint, Jobs & output, ValType & idleTime, bool useT);
 	size_t MaxJobCount();
